Audio_Python/requirements.txt
pyaudio
wave
keyboard
openai-whisper
pyserial
RPi.GPIO